requirements.txt
pytesseract
fastapi
uvicorn
sqlalchemy
python-multipart
aiofile
pillow

